% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTables.R
\name{extractTables}
\alias{extractTables}
\title{Extracts all xml <tables...> nodes into a list of data frames,
where each dataframe is the 'cell' nodes inside <table > content.}
\usage{
extractTables(fileName)
}
\arguments{
\item{fileName}{The file to convert}
}
\value{
List of data frames, where the key is a unique table name
constructed from the xml attributes 'Name' and 'Caption'
}
\description{
Extracts all xml <tables...> nodes into a list of data frames,
where each dataframe is the 'cell' nodes inside <table > content.
}
