% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTables.R, R/runApp.R
\name{runShiny}
\alias{runShiny}
\alias{runShiny}
\title{Runs the export as a shiny app, where the input file can be choosen
and the output is "table.xlsx:}
\usage{
runShiny(installDependencies = FALSE, ...)

runShiny(installDependencies = FALSE, ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Runs the export as a shiny app, where the input file can be choosen
and the output is "table.xlsx:

Run the shiny application
}
