% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTables.R
\name{writeToExcel}
\alias{writeToExcel}
\title{Writes the list of data frames to an excel file.
All cells have type 'text', so needed to be converted in Excel, if needed.}
\usage{
writeToExcel(dfs, file = "tables.xlxs")
}
\arguments{
\item{dfs}{A list of data frames}

\item{file}{The xlsx file to write the extracted tables to}
}
\description{
Writes the list of data frames to an excel file.
All cells have type 'text', so needed to be converted in Excel, if needed.
}

