% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTables.R
\name{runShiny}
\alias{runShiny}
\title{Runs the export as a shiny app, where the input file can be choosen
and the output is "table.xlsx:}
\usage{
runShiny()
}
\description{
Runs the export as a shiny app, where the input file can be choosen
and the output is "table.xlsx:
}

